requirements.txt
cycler==0.10.0
future==0.16.0
iso8601==0.1.12
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.15.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
# pyserial needs to be hard coded
pyserial==3.4